refactor: move type definitions to a ts file - Rating, Button, DatePicker
