+++ src/components/Button/index.ts
@@ -1,1 +1,2 @@
-export { default as Button, type ButtonProps } from './Button.vue'
+export { default as Button } from './Button.vue'
+export type { ButtonProps } from './types'
