+++ src/components/Rating/types.ts
@@ -1,0 +1,7 @@
+export interface RatingProps {
+	modelValue?: number
+	rating_from?: number
+	label?: string
+	readonly?: boolean
+	size?: 'sm' | 'md' | 'lg' | 'xl'
+}
